Moving discovery related interfaces into appropriate location
