+++ src/app/logs/duck/sagas.ts
@@ -4,8 +4,7 @@ import { flatten } from 'lodash';
 import { LogActions, LogActionTypes } from './actions';
 import Q from 'q';
 import { IDiscoveryClient } from '../../../client/discoveryClient';
-import { PlanPodReportDiscovery } from '../../../client/resources/discovery';
-import { IPlanReport, IPodLogSource, IPlanLogSources, ClusterKind } from '../../../client/resources/convension';
+import { PlanPodReportDiscovery, ClusterKind, IPlanLogSources, IPlanReport } from '../../../client/resources/discovery';
 import JSZip from 'jszip';
 import utils from '../../common/duck/utils';
 import { handleCertError } from './utils';
