+++ src/client/resources/common.ts
@@ -1,5 +1,3 @@
-import moment = require("moment");
-
 export type KubeResource = NamespacedResource | ClusterResource;
 
 export interface IKubeResource {
