+++ src/client/resources/discovery.ts
@@ -1,6 +1,26 @@
 import { NamedDiscoveryResource, DiscoveryResource, IDiscoveryParameters } from './common';
 import { IDiscoveryClient } from '../discoveryClient';
-import { IPlanReport } from './convension';
+
+export enum ClusterKind {
+  source = 'source',
+  destination = 'destination',
+  controller = 'controller',
+}
+
+export interface IPodLogSource {
+  name: string;
+  namespace: string;
+  log: string;
+}
+
+export type IPlanLogSources = {
+  [key in ClusterKind]: IPodLogSource[];
+};
+
+export interface IPlanReport extends IPlanLogSources {
+  namespace: string;
+  name: string;
+}
 
 export class NamespaceDiscovery extends DiscoveryResource {
   constructor(cluster: string, params: IDiscoveryParameters = {}) {
